feat: Create seperate SpeedControl component
Changes:
- Move speed control from MainWindow to a new SpeedControl component in the utils package
- Add SpeedControl to the MainWindow and DockingWindow
- Add a label showing time ratio in SpeedControl
- Ensure SpeedControl is unregistered when the window is closed

Ref: #1808

+++ mars-sim-core/src/main/java/com/mars_sim/core/time/ClockListener.java
@@ -15,7 +15,7 @@ public interface ClockListener {
      * Desired Speed of the clock has changed
      * @param desiredTR Requested time ratio.
      */
-    void desiredTimeRatioChange(double desiredTR);
+    void desiredTimeRatioChange(int desiredTR);
 
 	/**
 	 * Change the pause state of the clock.
